LPS-69255 Rename SubscriptionConstants class
