+++ modules/apps/collaboration/subscription/subscription-api/src/main/java/com/liferay/subscription/model/SubscriptionConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.subscription.constants;
+package com.liferay.subscription.model;
 
 /**
  * @author Alexander Chow
